move internals of generator bin to support crate
