+++ axum-on-rails/src/cli/mod.rs
@@ -1,1 +1,2 @@
 pub mod db;
+pub mod generate;
